Extract annotations-agnostic code in its own trait
