+++ lib/Doctrine/Persistence/Mapping/MappingException.php
@@ -37,6 +37,17 @@ class MappingException extends Exception
     }
 
     /**
+     * @param class-string $driverClassName
+     */
+    public static function pathRequiredForDriver(string $driverClassName): self
+    {
+        return new self(sprintf(
+            'Specifying the paths to your entities is required when using %s to retrieve all class names.',
+            $driverClassName
+        ));
+    }
+
+    /**
      * @param string|null $path
      *
      * @return self
