+++ tests/Doctrine/Tests/Persistence/Mapping/_files/colocated/TestClass.php
